排班：节目表和人员表为多对多关系，加了一张关系表进行关联；后端逻辑有待完善

+++ src/main/java/com/microsoft/schedule_tool/service/impl/schedule/ProgramEmployeeServiceImpl.java
@@ -1,6 +1,8 @@
 package com.microsoft.schedule_tool.service.impl.schedule;
 
+import com.microsoft.schedule_tool.dao.schedule.ProgramAndEmployeeRepository;
 import com.microsoft.schedule_tool.dao.schedule.ProgramEmployeeRepository;
+import com.microsoft.schedule_tool.entity.schedule.ProgramAndEmployee;
 import com.microsoft.schedule_tool.entity.schedule.ProgramEmployee;
 import com.microsoft.schedule_tool.exception.schedule.ProgramEmployeeException;
 import com.microsoft.schedule_tool.service.schedule.ProgramEmployeeService;
@@ -9,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -22,37 +25,68 @@ public class ProgramEmployeeServiceImpl implements ProgramEmployeeService {
     @Autowired
     private ProgramEmployeeRepository mProgramEmployeeRepository;
 
+    @Autowired
+    private ProgramAndEmployeeRepository mProgramAndEmployeeRepository;
+
     @Override
     public List<ProgramEmployee> getAllProgramEmployeesByProgramId(Integer programId) {
-        return mProgramEmployeeRepository.findByProgramId(programId);
+        List<ProgramAndEmployee> programAndEmployeeList = mProgramAndEmployeeRepository.findByProgramId(programId);
+
+        List<ProgramEmployee> programEmployeeList = new ArrayList<>();
+
+        for (ProgramAndEmployee programAndEmployee : programAndEmployeeList) {
+            Optional<ProgramEmployee> programEmployeeOptional = mProgramEmployeeRepository.findById(programAndEmployee.getEmployeeId());
+            if (programEmployeeOptional.isPresent()) {
+                ProgramEmployee employee = programEmployeeOptional.get();
+                programEmployeeList.add(employee);
+            }
+        }
+
+        return programEmployeeList;
     }
 
     @Transactional
     @Override
-    public ProgramEmployee saveProgramEmployee(ProgramEmployee programEmployee) {
+    public ProgramEmployee saveProgramEmployee(ProgramEmployee programEmployee, Integer programId) {
 
-        if(programEmployee==null||programEmployee.getName()==null||"".equals(programEmployee.getName())){
+        if (programEmployee == null || programEmployee.getName() == null || "".equals(programEmployee.getName())) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_NAME_NULL);
         }
 
-        if(programEmployee.getProgramId()==null){
-            throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_PROGRAM_ID_NULL);
-        }
+        Optional<ProgramEmployee> programEmployeeOptional = mProgramEmployeeRepository
+                .findByName(programEmployee.getName());
+        try {
+            if (programEmployeeOptional.isPresent()) {
+                ProgramEmployee employee = programEmployeeOptional.get();
+                Integer employeeId = employee.getId();
+                Optional<ProgramAndEmployee> programAndEmployeeOptional = mProgramAndEmployeeRepository
+                        .findByProgramIdAndEmployeeId(programId, employeeId);
+                if (!programAndEmployeeOptional.isPresent()) {
+                    ProgramAndEmployee programAndEmployee = new ProgramAndEmployee();
+                    programAndEmployee.setProgramId(programId);
+                    programAndEmployee.setEmployeeId(employeeId);
+                    mProgramAndEmployeeRepository.save(programAndEmployee);
+                } else {
+                    throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_EXIST);
+                }
 
-        Optional<ProgramEmployee> programEmployeeOptional=mProgramEmployeeRepository
-                .findByNameAndProgramId(programEmployee.getName(),programEmployee.getProgramId());
+                return employee;
+            } else {
 
-        if(programEmployeeOptional.isPresent()){
-            throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_EXIST);
-        }
-        try{
-            ProgramEmployee result=mProgramEmployeeRepository.save(programEmployee);
-            if(result!=null){
-                return result;
-            }else{
-                throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_SAVE_FAIL);
+                ProgramEmployee result = mProgramEmployeeRepository.save(programEmployee);
+                if (result != null) {
+                    Integer employeeId = result.getId();
+                    ProgramAndEmployee programAndEmployee = new ProgramAndEmployee();
+                    programAndEmployee.setProgramId(programId);
+                    programAndEmployee.setEmployeeId(employeeId);
+                    mProgramAndEmployeeRepository.save(programAndEmployee);
+                    return result;
+                } else {
+                    throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_SAVE_FAIL);
+                }
             }
-        }catch (Exception e){
+
+        } catch (Exception e) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_SAVE_FAIL);
         }
     }
@@ -61,43 +95,53 @@ public class ProgramEmployeeServiceImpl implements ProgramEmployeeService {
     @Override
     public ProgramEmployee updateProgramEmployee(Integer id, String name) {
 
-        if(!mProgramEmployeeRepository.findById(id).isPresent()){
+        if (!mProgramEmployeeRepository.findById(id).isPresent()) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_ID_NOT_EXIST);
         }
 
-        ProgramEmployee programEmployee=mProgramEmployeeRepository.findById(id).get();
+        ProgramEmployee programEmployee = mProgramEmployeeRepository.findById(id).get();
 
-        if(mProgramEmployeeRepository.findByNameAndProgramId(name,programEmployee.getProgramId()).isPresent()){
+        if (mProgramEmployeeRepository.findByName(name).isPresent()) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_EXIST);
         }
 
         programEmployee.setName(name);
 
         try {
-            ProgramEmployee result=mProgramEmployeeRepository.save(programEmployee);
-            if(result!=null){
+            ProgramEmployee result = mProgramEmployeeRepository.save(programEmployee);
+            if (result != null) {
                 return result;
-            }else{
+            } else {
                 throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_UPDATE_FAIL);
             }
-        }catch (Exception e){
+        } catch (Exception e) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_UPDATE_FAIL);
         }
     }
 
+    @Transactional
     @Override
-    public boolean deleteProgramEmployee(Integer id) {
+    public boolean deleteProgramEmployee(Integer id, Integer programId) {
 
-        if(!mProgramEmployeeRepository.findById(id).isPresent()){
+        if (!mProgramEmployeeRepository.findById(id).isPresent()) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_ID_NOT_EXIST);
         }
 
-        mProgramEmployeeRepository.deleteById(id);
+        mProgramAndEmployeeRepository.deleteByProgramIdAndEmployeeId(programId, id);
 
-        if(mProgramEmployeeRepository.findById(id).isPresent()){
+        if (mProgramAndEmployeeRepository.findByProgramIdAndEmployeeId(programId, id).isPresent()) {
             throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_DELETE_FAIL);
-        }else{
-            return true;
         }
+
+        if (mProgramAndEmployeeRepository.findByEmployeeId(id).size() == 0) {
+            mProgramEmployeeRepository.deleteById(id);
+
+            if (mProgramEmployeeRepository.findById(id).isPresent()) {
+                throw new ProgramEmployeeException(ResultEnum.PROGRAM_EMPLOYEE_DELETE_FAIL);
+            }
+
+        }
+
+        return true;
     }
 }
